Now we pass the '-l gi' (= scan-switching) iteration consistency assertion.
The order of operations in the Soft-DFS and Best-First-Search ("A-star")
scans was changed to accomodate for the iteration direct-incrementation
consistency.

Individual changes:

scans.c : made the traces run only if the FCS_TRACE environment variable
is defined, and fixed the problems.

t/t/intermediate-iterations-are-in-order.t : removed the $TODO from the 
"-l gi" assertion because it is now successful.


git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/trunk@1213 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a