Solved the problems with using the -opt flag after recycling instances.
This was exhibited in solving ranges by the range solver.

Changes:

t/t/range-parallel-solver.t: added a test to test for a large range of
deals with "-l cool-jives -opt". This failed before this commit.

    - Also removed several unused "use" statements.

t/t/valgrind.t: added a test to test for "1 2 1 -opt". This failed before
this commit.

t/t/intermediate-iterations-are-in-order.t: testing that the piped open
closes successfully.

freecell.c: now assigning scans_synergy in fc_solve_sfs_check_state_end()
to the right value. Previously - it was uninitialised. valgrind did not
like it.

intrface.c: now recycling instance->optimization_thread instead of destroying
it and re-initialising it. (Because it consumed the next_soft_thread_id's.

    - added a flag ->in_optimization_thread to indicate if the optimization
    thread is active or not. Using it instead of ->optimization_thread where
    apprpriate.

    - Now resetting the BFS queue of the soft_thread's.


git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/trunk@1493 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a