Start implementing meta_alloc.c and meta_alloc.h.
This aims to be a compact allocator scheme based on a meta allocator that 
will serve several allocators, by giving them already allocated pages from
a recycle bin. After freeing a child allocator, it will add all its allocated
buffers to the recycle bin for use by future allocations.

Currently, there is no meta-allocator - but the allocators were converted to
use a linked list instead of an array, which is another step for reducing
memory fragmentation, as well as the plan for the meta allocator.


git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/fc-solve/trunk@4395 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a