Fix the counting of the irreversible moves.
What happened was that when moving a card from a false parent directly
to the foundation, it was counted as only one irreversible move.
However, when the same is done in two steps: one to move the card to
a parent card, freecell or empty stack and the second to move it to the
foundation, it was counted as two moves.

As a result, this change takes this into account.