enable_freecells_to_empty_columns_pruning .
This is a prune parameter for that, that is mentioned in the SPEC
here:

http://groups.yahoo.com/neo/groups/fc-solve-discuss/conversations/topics/1358

Quoting from it:

<<<<<
Hi all,

today while taking a walk and thinking about my propopsed SPEC for the
split-solver for fully-connected-components (see
http://groups.yahoo.com/neo/groups/fc-solve-discuss/conversations/topics/1357 )
I came up with an interesting idea for a prune that may improve the runtime
performance/memory consumption of the solvers. I'd like to share it here and
ask for input.

What I think is that if a card is in the freecells, it should not be moved to
an empty column *unless* that move will be immediately be followed by a
sequence of cards (or a sub-sequence of them) moving on top of the card. That's
because the maximal sequence move increases better with the addition of
empty columns than it does with the addition of cards, so there is no point to
preferring to have many freecells over the same amount of empty columns.

So if 6H is in a freecell, then the only kosher moves are:

1. Moving it to the hearts foundation (naturally).

2. Moving it on top of a 7S or 7C.

3. Moving it on an empty column and immediately moving 5C or 5S on top of it.

Does anyone have any comments on that? This thing excites me so much that I'm
going to start working on it in a new git branch (which I can revert in case
there's something missing in my reasoning).

Regards and thanks in advance,

Shlomi Fish
>>>>>