Update NEWS.txt for the changes since last rel.
The upcoming "conspiracy-theory" preset is named so in honour of Purim
and the scroll of Esther. See:

* http://shlomif.livejournal.com/62243.html

* https://en.wikipedia.org/wiki/Purim

It was today. Anyway, it will be based on the recent scan from the
fc-solve/scripts/TEST_OPTIMIZATIONS directory.

I want to do this release now because it's been a long time since the
last major release and it's "Release early - release often".