Implement the set.c/set.h open addr hash table.
See https://github.com/anholt/hash_table . Modified somewhat to get it
to build.

The upside is that the tests and especially the valgrind tests appear to
run much faster (about 17 seconds instead of about 24s). Maybe that's a
system-wide improvement that affects the old code too.

The downside is that the benchmarks are quite slower - about 8.3s.