Enforce signed chars globally.
This fixes the tests (and the production code) on ARM Linux. I was
tipped regarding Scaleway which provides ARM Linux servers at an affordable
price which allowed me to build and test Freecell Solver there. Thanks
also to the Mageia build system for catching this problem on ARM in the
first place.

We need to do a more failproof conversion of "char"s to "signed char"s
in the code, where appropriate, but this is a quick workaround.